Remove accidental recursive symlinking

+++ scripts/install_configs.sh
@@ -9,18 +9,18 @@ dotfiles_config_dir="${user_home_dir}/dotfiles/config"
 mkdir -p "$config_dir"
 
 # Symlinked Directories
-ln -s "${dotfiles_config_dir}/bat" "${config_dir}/bat"
-ln -s "${dotfiles_config_dir}/helix" "${config_dir}/helix"
-ln -s "${dotfiles_config_dir}/joshuto" "${config_dir}/joshuto"
-ln -s "${dotfiles_config_dir}/pet" "${config_dir}/pet"
-ln -s "${dotfiles_config_dir}/ranger" "${config_dir}/ranger"
-ln -s "${dotfiles_config_dir}/zellij" "${config_dir}/zellij"
+if [ ! -e "${config_dir}/bat" ]; then ln -s "${dotfiles_config_dir}/bat" "${config_dir}/bat"; fi
+if [ ! -e "${config_dir}/helix" ]; then ln -s "${dotfiles_config_dir}/helix" "${config_dir}/helix"; fi
+if [ ! -e "${config_dir}/joshuto" ]; then ln -s "${dotfiles_config_dir}/joshuto" "${config_dir}/joshuto"; fi
+if [ ! -e "${config_dir}/pet" ]; then ln -s "${dotfiles_config_dir}/pet" "${config_dir}/pet"; fi
+if [ ! -e "${config_dir}/ranger" ]; then ln -s "${dotfiles_config_dir}/ranger" "${config_dir}/ranger"; fi
+if [ ! -e "${config_dir}/zellij" ]; then ln -s "${dotfiles_config_dir}/zellij" "${config_dir}/zellij"; fi
 
 # Symlinked Files
-ln -s "${dotfiles_config_dir}/.bash_aliases" "${user_home_dir}/.bash_aliases"
-ln -s "${dotfiles_config_dir}/.inputrc" "${user_home_dir}/.inputrc"
-ln -s "${dotfiles_config_dir}/lazygit/config.yml" "${config_dir}/lazygit/config.yml"
-ln -s "${dotfiles_config_dir}/starship.toml" "${config_dir}/starship.toml"
+if [ ! -e "${user_home_dir}/.bash_aliases" ]; then ln -s "${dotfiles_config_dir}/.bash_aliases" "${user_home_dir}/.bash_aliases"; fi
+if [ ! -e "${user_home_dir}/.inputrc" ]; then ln -s "${dotfiles_config_dir}/.inputrc" "${user_home_dir}/.inputrc"; fi
+if [ ! -e "${config_dir}/lazygit/config.yml" ]; then ln -s "${dotfiles_config_dir}/lazygit/config.yml" "${config_dir}/lazygit/config.yml"; fi
+if [ ! -e "${config_dir}/starship.toml" ]; then ln -s "${dotfiles_config_dir}/starship.toml" "${config_dir}/starship.toml"; fi
 
 # PUDB can't read symlinks, so I need to copy/overwrite instead
 mkdir -p ~/.config/pudb
